Fix ctrl c
